re_build_tools
Module hashing

Functions

compute_crate_hash

Given a crate name, computes the sha256 hash of its source code (ordered by filename) and returns an hexadecimal string for it.

compute_dir_filtered_hash

Given a directory path, computes the sha256 hash of the accumulated contents of all of its files (ordered by filename) except those failing a custom filter, and returns an hexadecimal string for it.

compute_dir_hash

Given a directory path, computes the sha256 hash of the accumulated contents of all of its files (ordered by filename), and returns an hexadecimal string for it.

compute_file_hash

Given a file path, computes the sha256 hash of its contents and returns an hexadecimal string for it.

compute_strings_hash

Given a bunch of strings, computes the sha256 hash of their contents (in the order they were passed in) and returns an hexadecimal string for it.

encode_hex 🔒

iter_dir

Recursively walks the directory at path in filename order.

iter_dir_filtered

Recursively walks the directory at path in filename order with an arbitrary filter.

read_versioning_hash

Reads back a versioning hash that was written with write_versioning_hash.

write_versioning_hash

Writes the given hash at the specified path.