re_build_tools::rebuild_detector
Function rebuild_if_crate_changed

pub fn rebuild_if_crate_changed(pkg_name: &str)

Call from build.rs to trigger a rebuild whenever any source file of the given package or any of its dependencies changes, recursively.

This will work even if the package depends on crates that are outside of the workspace, included with path = …

However, this is a complex beast, and may have bugs in it. Maybe it is even causing spurious re-compiles (https://github.com/rerun-io/rerun/issues/3266).