re_build_tools::rebuild_detector
Struct Packages

pub struct Packages<'a> {
    pkgs: HashMap<&'a str, &'a Package>,
}

Fields
pkgs: HashMap<&'a str, &'a Package>
Implementations

impl<'a> Packages<'a>

pub fn from_metadata(metadata: &'a Metadata) -> Self

pub fn track_implicit_dep( &self, pkg_name: &str, files_to_watch: &mut HashSet<PathBuf>, )

Tracks an implicit dependency of the given name.

This will generate all the appropriate cargo:rerun-if-changed clauses so that package pkg_name as well as all of it direct and indirect dependencies are properly tracked whether they are remote, in-workspace, or locally patched.

fn track_patched_deps( &self, tracked: &mut HashSet<PackageId>, root: &Package, files_to_watch: &mut HashSet<PathBuf>, )

Recursively walk the tree of dependencies of the given root package, making sure to track all potentially modified, locally patched dependencies.