Crate re_build_tools

This crate is to be used from build.rs build scripts.

Use this crate together with the re_build_info crate.

Modules

hashing 🔒

rebuild_detector 🔒

Statics

OUTPUT_CARGO_BUILD_INSTRUCTIONS 🔒

Atomic bool indicating whether or not to print the cargo build instructions

Functions

compute_crate_hash

Given a crate name, computes the sha256 hash of its source code (ordered by filename) and returns an hexadecimal string for it.

compute_dir_filtered_hash

Given a directory path, computes the sha256 hash of the accumulated contents of all of its files (ordered by filename) except those failing a a custom filter, and returns an hexadecimal string for it.

compute_dir_hash

Given a directory path, computes the sha256 hash of the accumulated contents of all of its files (ordered by filename), and returns an hexadecimal string for it.

compute_file_hash

Given a file path, computes the sha256 hash of its contents and returns an hexadecimal string for it.

compute_strings_hash

Given a bunch of strings, computes the sha256 hash of their contents (in the order they were passed in) and returns an hexadecimal string for it.

export_env_vars

Call from the build.rs file of any crate you want to generate build info for.

get_and_track_env_var

Call from build.rs to trigger a rebuild whenever an environment variable changes.

git_branch 🔒

git_hash 🔒

git_path 🔒

From https://git-scm.com/docs/git-rev-parse:

is_tracked_env_var_set

Call from build.rs to trigger a rebuild whenever an environment variable changes, and returns true if that variable has been set to a truthy value.

iter_dir

Recursively walks the directory at path in filename order.

read_versioning_hash

Reads back a versioning hash that was written with write_versioning_hash.

rebuild_if_crate_changed

Call from build.rs to trigger a rebuild whenever any source file of the given package or any of its dependencies changes, recursively.

rerun_if_changed

Call from build.rs to trigger a rebuild whenever the file at path changes.

rerun_if_changed_glob

Call from build.rs to trigger a rebuild whenever any of the files identified by the given globbed path change.

rerun_if_changed_or_doesnt_exist

Call from build.rs to trigger a rebuild whenever the file at path changes, or it doesn’t exist.

run_command 🔒

rust_version 🔒

Returns (rustc, LLVM) versions.

set_env 🔒

set_output_cargo_build_instructions

Change whether or not this library should output cargo build instructions

should_output_cargo_build_instructions 🔒

Helper to check whether or not cargo build instructions should be printed.

write_file_if_necessary

Writes content to a file iff it differs from what’s already there.

write_versioning_hash

Writes the given hash at the specified path.