Module re_build_tools::rebuild_detector

Structs

Packages

Functions

get_and_track_env_var

Call from build.rs to trigger a rebuild whenever an environment variable changes.

is_tracked_env_var_set

Call from build.rs to trigger a rebuild whenever an environment variable changes, and returns true if that variable has been set to a truthy value.

rebuild_if_crate_changed

Call from build.rs to trigger a rebuild whenever any source file of the given package or any of its dependencies changes, recursively.

rerun_if_changed

Call from build.rs to trigger a rebuild whenever the file at path changes.

rerun_if_changed_glob

Call from build.rs to trigger a rebuild whenever any of the files identified by the given globbed path change.

rerun_if_changed_or_doesnt_exist

Call from build.rs to trigger a rebuild whenever the file at path changes, or it doesn’t exist.

write_file_if_necessary

Writes content to a file iff it differs from what’s already there.